refactor: Move section chunking to store
